+++ static/console/static/js/Bar-BtchnKqH.js
@@ -1,1 +1,0 @@
-import{_ as o}from"./Bar.vue_vue_type_script_setup_true_lang-DRB7NiwK.js";import"./app-Ck2FkEtM.js";import"./index-D-SOXmnp.js";export{o as default};